Add files via upload
This is an optimization project I completed in my Mathematics of Finance class. I found the perfect solution with this spread sheet. Here is the description: 

Project: “Plus one, minus one” game
A casino offers the following game: the customer is given a well-shuffled deck of N = 52
cards. The cards are turned over one-by-one and the customer gets $1 for each black card
and -$1 for each red card. The customer can stop at any time, collecting their winnings
(if the customer is losing, there is no reason to stop: once the whole deck is used up, the
customer is automatically back to $0). Design the procedure and write code to
(1) Compute the risk-neutral value of the game.
(2) Determine the optimal stopping strategy.
The risk-neutral value is the expected winnings maximized over all valid strategies. A
valid (deterministic) strategy is a function that decides stopping (0 or 1) based only on
the information available at the time: number of cards left in the deck and the current
winnings balance.